Remove `bindEmitter` call
After doing research on what `bindEmitter` does, I could not see the
benefits of using it. At the time the hooks run, the transaction is already
over. There seem to be no reason to preserve store context for userland
use.

All current tests pass. There is no mention of preserved context in the
README, therefore this should be treated as internal change.

This change is motivated by the fact AsyncLocalStorage does not support
`bindEmitter` out of the box and without understanding the end goal it
is not feasible to properly reimplement it.

If at any point in the future a viable use-case is provided, the bindEmitter
can easily be restored.